+++ modules/db/index.js
@@ -2,7 +2,7 @@ const init = (osseus) => {
   this.osseus = osseus
   return new Promise((resolve, reject) => {
     this.osseus.db_models = {
-      user: require('./models/User')(this.osseus)
+      funding: require('./models/Funding')(this.osseus)
     }
     osseus.logger.info(`DB ready`)
     return resolve()
